Added time and mask methods to Image class
These are currently needed for the interpolation.  I need to check to see if the time image really needs to be built from the ET image or whether we can use a simpler product (like NDVI or one of the source bands) but for now I am leaving as is to match the other model code.