PEP8 - Standardizing indentation spacing
I went with 4 spaces per tab to be consistent but I would be happy to change this to 2 per tab if that is preferred.

I wasn't sure the best way to handle the expression calls so I tried to be consistent with the existing code, but it would be good to decide whether to start the dictionary inputs on a new line or at the end of the existing.